Collection node ready for testing

+++ index.js
@@ -50,7 +50,7 @@ client.on("ready", () => {
 });
 
 client.on("message", (message) => {
-    if (message.author.bot) return;
+    if (message.author.bot || message.guild === null) return;
     if (!Helpers.isCommand(message, global.config.PREFIX)) {
         if (!Helpers.canSpeakInChannel(message, notify = true)) message.delete();
         return;
